refactor(diff): remove churn utils file and handle empty input
