+++ git-productivity-analyzer/src/sdk/churn/mod.rs
@@ -1,4 +1,3 @@
 mod analyzer;
-mod diff_utils;
 
 pub use analyzer::Options;
